pyproject.toml
[project]
name = "sehatmandmaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tayyaba", email = "tayyabashahbaz2020@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
sehatmandmaa = "sehatmandmaa:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
